refactor: add folders for resolvers
