+++ src/groups/groups.module.ts
@@ -3,10 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { EventsModule } from '../events/events.module';
 import { Image } from '../images/models/image.model';
 import { UsersModule } from '../users/users.module';
-import { GroupConfigsResolver } from './group-configs.resolver';
-import { GroupMemberRequestsResolver } from './group-member-requests.resolver';
+import { GroupConfigsResolver } from './resolvers/group-configs.resolver';
+import { GroupMemberRequestsResolver } from './resolvers/group-member-requests.resolver';
 import { GroupRolesModule } from './group-roles/group-roles.module';
-import { GroupsResolver } from './groups.resolver';
+import { GroupsResolver } from './resolvers/groups.resolver';
 import { GroupsService } from './groups.service';
 import { GroupConfig } from './models/group-config.model';
 import { GroupMemberRequest } from './models/group-member-request.model';
