+++ src/groups/resolvers/groups.resolver.ts
@@ -8,27 +8,27 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { CurrentUser } from '../auth/decorators/current-user.decorator';
-import { FeedItemsConnection } from '../common/models/feed-items.connection';
-import { Dataloaders } from '../dataloader/dataloader.types';
-import { EventsService } from '../events/events.service';
-import { Event } from '../events/models/event.model';
-import { EventTimeFrame } from '../events/models/events.input';
-import { User } from '../users/models/user.model';
-import { GroupRolesService } from './group-roles/group-roles.service';
-import { GroupPermissions } from './group-roles/models/group-permissions.type';
-import { GroupRole } from './group-roles/models/group-role.model';
-import { GroupPrivacy } from './groups.constants';
-import { GroupsService } from './groups.service';
-import { CreateGroupInput } from './models/create-group.input';
-import { CreateGroupPayload } from './models/create-group.payload';
-import { GroupConfig } from './models/group-config.model';
-import { GroupMemberRequest } from './models/group-member-request.model';
-import { Group } from './models/group.model';
-import { GroupsConnection } from './models/groups.connection';
-import { PublicGroupsConnection } from './models/public-groups.connection';
-import { UpdateGroupInput } from './models/update-group.input';
-import { UpdateGroupPayload } from './models/update-group.payload';
+import { CurrentUser } from '../../auth/decorators/current-user.decorator';
+import { FeedItemsConnection } from '../../common/models/feed-items.connection';
+import { Dataloaders } from '../../dataloader/dataloader.types';
+import { EventsService } from '../../events/events.service';
+import { Event } from '../../events/models/event.model';
+import { EventTimeFrame } from '../../events/models/events.input';
+import { User } from '../../users/models/user.model';
+import { GroupRolesService } from '../group-roles/group-roles.service';
+import { GroupPermissions } from '../group-roles/models/group-permissions.type';
+import { GroupRole } from '../group-roles/models/group-role.model';
+import { GroupPrivacy } from '../groups.constants';
+import { GroupsService } from '../groups.service';
+import { CreateGroupInput } from '../models/create-group.input';
+import { CreateGroupPayload } from '../models/create-group.payload';
+import { GroupConfig } from '../models/group-config.model';
+import { GroupMemberRequest } from '../models/group-member-request.model';
+import { Group } from '../models/group.model';
+import { GroupsConnection } from '../models/groups.connection';
+import { PublicGroupsConnection } from '../models/public-groups.connection';
+import { UpdateGroupInput } from '../models/update-group.input';
+import { UpdateGroupPayload } from '../models/update-group.payload';
 
 @Resolver(() => Group)
 export class GroupsResolver {
