+++ src/proposals/proposal-actions/proposal-actions.module.ts
@@ -13,11 +13,11 @@ import { ProposalActionPermission } from './models/proposal-action-permission.mo
 import { ProposalActionRoleMember } from './models/proposal-action-role-member.model';
 import { ProposalActionRole } from './models/proposal-action-role.model';
 import { ProposalAction } from './models/proposal-action.model';
-import { ProposalActionEventHostsResolver } from './proposal-action-event-hosts.resolver';
-import { ProposalActionEventsResolver } from './proposal-action-events.resolver';
-import { ProposalActionGroupConfigsResolver } from './proposal-action-group-configs.resolver';
-import { ProposalActionRolesResolver } from './proposal-action-roles.resolver';
-import { ProposalActionsResolver } from './proposal-actions.resolver';
+import { ProposalActionEventHostsResolver } from './resolvers/proposal-action-event-hosts.resolver';
+import { ProposalActionEventsResolver } from './resolvers/proposal-action-events.resolver';
+import { ProposalActionGroupConfigsResolver } from './resolvers/proposal-action-group-configs.resolver';
+import { ProposalActionRolesResolver } from './resolvers/proposal-action-roles.resolver';
+import { ProposalActionsResolver } from './resolvers/proposal-actions.resolver';
 import { ProposalActionsService } from './proposal-actions.service';
 
 @Module({
