+++ src/proposals/proposal-actions/resolvers/proposal-action-event-hosts.resolver.ts
@@ -1,7 +1,7 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { User } from '../../users/models/user.model';
-import { ProposalActionEventHost } from './models/proposal-action-event-host.model';
-import { ProposalActionsService } from './proposal-actions.service';
+import { User } from '../../../users/models/user.model';
+import { ProposalActionEventHost } from '../models/proposal-action-event-host.model';
+import { ProposalActionsService } from '../proposal-actions.service';
 
 @Resolver(() => ProposalActionEventHost)
 export class ProposalActionEventHostsResolver {
