+++ src/proposals/proposal-actions/resolvers/proposal-action-group-configs.resolver.ts
@@ -1,7 +1,7 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { ProposalAction } from './models/proposal-action.model';
-import { ProposalActionsService } from './proposal-actions.service';
-import { ProposalActionGroupConfig } from './models/proposal-action-group-config.model';
+import { ProposalAction } from '../models/proposal-action.model';
+import { ProposalActionsService } from '../proposal-actions.service';
+import { ProposalActionGroupConfig } from '../models/proposal-action-group-config.model';
 
 @Resolver(() => ProposalActionGroupConfig)
 export class ProposalActionGroupConfigsResolver {
