+++ src/proposals/proposal-actions/resolvers/proposal-action-roles.resolver.ts
@@ -1,10 +1,10 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { GroupRolesService } from '../../groups/group-roles/group-roles.service';
-import { GroupRole } from '../../groups/group-roles/models/group-role.model';
-import { ProposalActionPermission } from './models/proposal-action-permission.model';
-import { ProposalActionRoleMember } from './models/proposal-action-role-member.model';
-import { ProposalActionRole } from './models/proposal-action-role.model';
-import { ProposalActionsService } from './proposal-actions.service';
+import { GroupRolesService } from '../../../groups/group-roles/group-roles.service';
+import { GroupRole } from '../../../groups/group-roles/models/group-role.model';
+import { ProposalActionPermission } from '../models/proposal-action-permission.model';
+import { ProposalActionRoleMember } from '../models/proposal-action-role-member.model';
+import { ProposalActionRole } from '../models/proposal-action-role.model';
+import { ProposalActionsService } from '../proposal-actions.service';
 
 @Resolver(() => ProposalActionRole)
 export class ProposalActionRolesResolver {
